pyproject.toml
[project]
name = "simple-crawler"
version = "0.1.0"
dependencies = [
    "requests",
    "beautifulsoup4",
    "PyYAML",
]
